pyproject.toml
[tool.aerich]
tortoise_orm = "src.db.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = ["black>=23.12.1", "faker>=22.5.0"]

[tool.pdm.build]
includes = []

[tool.pdm.scripts]
dev = "uvicorn src.main:app --host 0.0.0.0 --port 4000 --reload"
pro = "gunicorn -c gunicorn.conf.py src.main:app"
aerich_init = "aerich init -t src.core.db.TORTOISE_ORM"
init_db = "aerich init-db"
migrate_db = "aerich migrate"
upgrade_db = "aerich upgrade"
test = "pytest tests -sv"
test_add_user = "pytest tests -sv -m add_user"

[tool.pytest.ini_options]
# pytest配置
addopts = "-p no:warnings"
markers = ["add_user: add fake user"]


[project]
name = ""
version = ""
description = ""
authors = [{ name = "RicarTang", email = "357965604@qq.com" }]
dependencies = [
    "pydantic==2.9.2",
    "tortoise-orm[asyncmy]<1.0.0,>=0.20.0",
    "gunicorn<22.0.0,>=21.2.0",
    "passlib<2.0.0,>=1.7.4",
    "python-jose<4.0.0,>=3.3.0",
    "apscheduler<4.0.0,>=3.10.4",
    "aerich<1.0.0,>=0.7.2",
    "openpyxl<4.0.0,>=3.1.2",
    "aiofiles<24.0.0,>=23.2.1",
    "pyyaml<7.0.0,>=6.0.1",
    "loguru<1.0.0,>=0.7.2",
    "pytest<8.0.0,>=7.4.3",
    "sse-starlette<2.0.0,>=1.8.2",
    "pydantic-settings<3.0.0,>=2.1.0",
    "python-multipart<1.0.0,>=0.0.6",
    "fastapi<1.0.0,>=0.115",
    "uvicorn[standard]<1.0.0.0,>=0.24.0.post1",
    "celery>=5.3.6",
    "cryptography>=41.0.7",
    "pymysql>=1.1.0",
    "pytest-asyncio>=0.23.2",
    "allure-pytest>=2.13.2",
    "redis[hiredis]>=5.0.1",
    "fastapi-cache2>=0.2.1",
    "pendulum>=2.1.2",
    "httpx>=0.25.2",
    "pytest-assume>=2.4.3",
    "asgi-lifespan>=2.1.0",
    "sqlalchemy>=2.0.25",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
